added test file

+++ application/routes.py
@@ -4,7 +4,7 @@ from flask import request, redirect, url_for
 from flask import render_template
 from flask_wtf import FlaskForm
 from wtforms import StringField, SubmitField
-from application.forms import NameForm, DelForm, SelectFieldForm, SinForm, MoodleForm
+from application.forms import NameForm, DelForm, SinForm, MoodleForm, SelectFieldForm
 
 #****************FUNCTION****select_List*********************************************
 def select_list(passdata):
@@ -24,6 +24,14 @@ def select_list(passdata):
              return redirect(url_for('readstudent'))
         elif passdata.list.data=='Home Page':
              return redirect(url_for('start'))  
+#*******************************************************************
+
+# def exitbut():
+          
+#           if  request.method == 'POST':
+#              return redirect(url_for('start'))
+#           else:
+#               0
 
 
 #*****************************************************************************************
@@ -33,7 +41,8 @@ def select_list(passdata):
 def start():
     pyoptionform= SelectFieldForm()
    # return redirect(url_for('read'))
-    if request.method == 'POST':
+    #if request.method == 'POST':
+    if pyoptionform.validate_on_submit():
        return select_list(pyoptionform)
     return render_template('home.html', jioptionform=pyoptionform)
 
@@ -44,7 +53,8 @@ def start():
 def read_all_student():
     pyall_students = student.query.all()
     pyoptionform= SelectFieldForm()
-    if request.method == 'POST':
+   # if request.method == 'POST':
+    if pyoptionform.validate_on_submit():
         return select_list(pyoptionform)
     return render_template('allstudent.html', ji_all_student=pyall_students, jioptionform=pyoptionform)
 
@@ -56,7 +66,8 @@ def read_all_student():
 def read_all_moodle():
     pyall_moodle = moodle.query.all()
     pyoptionform1= SelectFieldForm()
-    if request.method == 'POST':
+    #if request.method == 'POST':
+    if pyoptionform1.validate_on_submit():
       return select_list(pyoptionform1)
     return render_template('allmoodle.html', ji_all_moodle= pyall_moodle, jioptionform1=pyoptionform1)
 
@@ -67,11 +78,20 @@ def read_all_moodle():
 @app.route('/regstudent', methods = ['GET','POST'])
 def register_student():
     pystudentform = NameForm()
-    if request.method == 'POST':
-        addstudent = student(student_id=pystudentform.Stud_id.data, surname=pystudentform.SName.data,firstName=pystudentform.FName.data )
-        schooldb.session.add(addstudent)
-        schooldb.session.commit()
-        return redirect(url_for('start'))
+    if     request.method == 'POST':
+             return redirect(url_for('start'))
+    else:
+    #  pystudentform = NameForm()
+      #    if request.method == 'POST':
+          if pystudentform.validate_on_submit():
+             addstudent = student(student_id=pystudentform.Stud_id.data, surname=pystudentform.SName.data,firstName=pystudentform.FName.data )
+             schooldb.session.add(addstudent)
+             schooldb.session.commit()
+             return redirect(url_for('start'))
+          else:
+               return redirect(url_for('start'))
+              #return render_template('error.html', jistudentform=pystudentform)
+    
     return render_template('registerstudent.html', jistudentform=pystudentform)
 
 
@@ -80,18 +100,25 @@ def register_student():
 #****************DELETE ENTRY FROM THE STUDENT TABLE*******************
 @app.route('/A', methods = ['GET','POST'])
 def delstudent():
-         pyall_students = student.query.all()
-         pydelstudentform=DelForm()
-         if request.method == 'POST':
-            delmoodle = moodle.query.filter_by(student_id=pydelstudentform.Stud_id.data).all()
-            for k in delmoodle:
-              schooldb.session.delete(k)            
-            schooldb.session.commit()
-            delstudent = student.query.filter_by(student_id=pydelstudentform.Stud_id.data).first()
-            schooldb.session.delete(delstudent)
-            schooldb.session.commit()
+        if  request.method == 'POST':
             return redirect(url_for('start'))
-         return render_template('deletebut.html', ji_all_student=pyall_students,jidelstudentform=pydelstudentform )
+        else:
+              # exitbut()
+            pyall_students = student.query.all()
+            pydelstudentform=DelForm()
+            #if request.method == 'POST':
+            if pydelstudentform.validate_on_submit():
+                delmoodle = moodle.query.filter_by(student_id=pydelstudentform.Stud_id.data).all()
+                for k in delmoodle:
+                   schooldb.session.delete(k)            
+                   schooldb.session.commit()
+                   delstudent = student.query.filter_by(student_id=pydelstudentform.Stud_id.data).first()
+                   schooldb.session.delete(delstudent)
+                   schooldb.session.commit()
+                return redirect(url_for('start'))
+        return render_template('deletebut.html', ji_all_student=pyall_students,jidelstudentform=pydelstudentform )
+        
+            
 
                                                                             
 #********************************************************************************************
@@ -100,11 +127,15 @@ def delstudent():
 
 @app.route('/studmod', methods=['GET','POST'])
 def readstudent():
-         pystudentform=SinForm()
-         if request.method == 'POST':
-          student_id=pystudentform.Stud_id.data
-          return redirect(url_for('readresult',student_id1= student_id))
-         return render_template('singlestud.html', jistudentform=pystudentform,)
+        # if  request.method == 'POST':
+        #      return redirect(url_for('start'))
+        # else:
+             pystudentform=SinForm()
+             #if request.method == 'POST':
+             if pystudentform.validate_on_submit():
+                   student_id=pystudentform.Stud_id.data
+                   return redirect(url_for('readresult',student_id1= student_id))
+             return render_template('singlestud.html', jistudentform=pystudentform,)
 
 @app.route('/studmod1/<student_id1>', methods=['GET','POST'])
 def readresult(student_id1):
@@ -123,11 +154,16 @@ def readresult(student_id1):
 
 @app.route('/studmodm', methods=['GET','POST'])
 def update():
-         pystudentform=SinForm()
-         if request.method == 'POST':
-          student_id=pystudentform.Stud_id.data
-          return redirect(url_for('register_student1',student_id1= student_id))
-         return render_template('singlestud.html', jistudentform=pystudentform,)
+        if  request.method == 'POST':
+            return redirect(url_for('start'))
+        else:
+           #  exitbut()
+             pystudentform=SinForm()
+            # if request.method == 'POST':
+             if pystudentform.validate_on_submit():
+               student_id=pystudentform.Stud_id.data
+               return redirect(url_for('register_student1',student_id1= student_id))
+        return render_template('singlestud.html', jistudentform=pystudentform,)
 
 @app.route('/studmod2/<student_id1>', methods=['GET','POST'])
 def register_student1(student_id1):
@@ -137,7 +173,8 @@ def register_student1(student_id1):
          studmoodle = moodle.query.filter_by(student_id=student_id1).all()
          pystudent = student.query.filter_by(student_id=student_id1).first()
          return render_template('update.html', ji_student=pystudent, ji_moodle= studmoodle, jioptionform= pyoptionform, jistudentform=pystudentform)
-    if request.method == 'POST':
+    #if request.method == 'POST':
+    if pyoptionform.validate_on_submit():
         studmoodle = moodle.query.filter_by(student_id=student_id1).all()
         for k in studmoodle:
             schooldb.session.delete(k)            
@@ -155,12 +192,17 @@ def register_student1(student_id1):
 
 @app.route('/regmoodle', methods = ['GET','POST'])
 def register_moodle():
-    pymodform = MoodleForm()
-    if request.method == 'POST':
-        addmoodle = moodle(student_id=pymodform.Stud_id.data, moodle_id=pymodform.Modcode.data, moodleName=pymodform.Subname.data )
-        schooldb.session.add(addmoodle)
-        schooldb.session.commit()
+    if  request.method == 'POST':
         return redirect(url_for('start'))
+    else:
+       # exitbut()
+        pymodform = MoodleForm()
+        #if request.method == 'POST':
+        if pymodform.validate_on_submit():
+           addmoodle = moodle(student_id=pymodform.Stud_id.data, moodle_id=pymodform.Modcode.data, moodleName=pymodform.Subname.data )
+           schooldb.session.add(addmoodle)
+           schooldb.session.commit()
+           return redirect(url_for('start'))
     return render_template('regmoodle.html', jimodform=pymodform)
